refactor: :art: Project Catalog Adjustment
