+++ artscription-ui/src/app/artscriptions/page.tsx
@@ -1,0 +1,9 @@
+export default async function Page() {
+    return (
+        <main>
+            <h1 className={` mb-4 text-xl md:text-2xl`}>
+                artscriptions
+            </h1>
+        </main>
+    );
+}
