+++ artscription-ui/src/app/layout.tsx
@@ -1,27 +1,23 @@
-import type { Metadata } from 'next'
-import { Inter } from 'next/font/google'
-import { Web3Modal } from "../context/Web3Modal";
-import './globals.css'
-
-const inter = Inter({ subsets: ['latin'] })
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
-
+import '@/ui/globals.css';
+import { Web3Modal } from '@/context/Web3Modal';
+import Navbar from '@/components/Navbar';
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
     <html lang="en">
-      <body className={inter.className}>
+      <body>
         <Web3Modal>
-          {children}
+          <div className="flex flex-col min-h-screen">
+            <Navbar></Navbar>
+            <div className='mt-20'>
+            {children}
+            </div>
+          </div>
         </Web3Modal>
       </body>
     </html>
-  )
+  );
 }
