+++ artscription-ui/src/app/lib/artscriptionDetail.js
