+++ artscription-ui/src/app/lib/proxy.js
