+++ artscription-ui/src/app/marketplace/page.tsx
@@ -1,0 +1,9 @@
+export default async function Page() {
+    return (
+        <main>
+            <h1 className={` mb-4 text-xl md:text-2xl`}>
+                marketplace
+            </h1>
+        </main>
+    );
+}
