+++ artscription-ui/src/app/page.tsx
@@ -1,0 +1,12 @@
+
+
+import { ConnectButton } from '@rainbow-me/rainbowkit';
+
+export default function Page() {
+  return (
+    <div>
+      <div className='h-[500px] bg-green-500'>6666</div>
+    </div>
+
+  );
+}
