+++ artscription-ui/src/public/favicon.ico
